Animate a Bezier curve between the two transport sites.
After a new location is picked, two random control points are also
picked, and a cubic Bezier curve is fitted between the starting and
ending transport locations. It consists of up to 50 points, and is
animated. They represent the player's molecules being disassembled,
flung across the universe, and reassembled at the new location.