One way portals (#6)
* Change how portals are created.

Now instead of cycling through a pattern of two-way portals, they are
randomly chosen from all 16 possible one- and two-way portal
combinations.

* Adjust the colors. Only one group is dependent on &background now.

* Draw a border of sorts on edges with active portals.

* Fix highlight group reference in statusline to new name.

* Bring back the hyperbolic tangent function with more reasonable limit.

The new asymptote is half as high as it used to be: l:cells/2. This is
more reasonable. This function also is continuous, not that anyone would
notice if they ever goto to level 64 with the min(...) formula.

* Add robot poo, and the ability to turn off the border.

Robot poo is the trail each robot leaves behind when it chases the
player down after pressing the F key.

The border uses Unicode characters, which may not show up well in some
fonts, so now there's an setting to turn them off.

* Fix default setting for g:robots_border to match the README.

* Display Robots Remaining as 'n of total' instead of just 'n'.

* Ensure borders are redrawn after transport, if departing/landing there.

* Better way to restore border wall after transport.

This change uses the same restoration logic that the s:Bezier function
uses - save the original characters in a List, and restore them when
done mucking about with the screen.

This fixes the space character showing up before the border is restored,
as well as the moving cursor.

* Fix player leaving robot poo behind when `F`inishing a round.

* Display Portal messages forever after their introduction.